fix(tokenizer): admit halfwidth katakana names to prefilter, split scan
- Fold halfwidth kana throughout so the character-name prefilter treats halfwidth spellings like any other; only unfoldable voiced marks in a candidate region bypass it, and kanji-then-halfwidth names (山ｶﾞｸ) now match too
- Split yomitan-scanning-helpers-script.ts into focused injected-script modules: kana, furigana, frequency, dictionary-classification, match-selection
- Extract shared vm-context test harness (yomitan-scan-test-harness.ts) out of yomitan-parser-runtime.test.ts and add yomitan-scan-runtime.test.ts
