+++ src/core/services/tokenizer/yomitan-parser-runtime.test.ts
@@ -3,7 +3,12 @@ import * as fs from 'fs';
 import * as os from 'os';
 import * as path from 'path';
 import test from 'node:test';
-import * as vm from 'node:vm';
+import {
+  countTermsFindLookups,
+  createDeps,
+  createScanDeps,
+  runInjectedYomitanScript,
+} from './yomitan-scan-test-harness';
 import {
   addYomitanNoteViaSearch,
   clearYomitanParserCachesForWindow,
@@ -17,330 +22,6 @@ import {
   syncYomitanDefaultAnkiServer,
   upsertYomitanDictionarySettings,
 } from './yomitan-parser-runtime';
-
-function createDeps(
-  executeJavaScript: (script: string) => Promise<unknown>,
-  options?: {
-    createYomitanExtensionWindow?: (pageName: string) => Promise<unknown>;
-  },
-) {
-  const parserWindow = {
-    isDestroyed: () => false,
-    webContents: {
-      executeJavaScript: async (script: string) => await executeJavaScript(script),
-    },
-  };
-
-  return {
-    getYomitanExt: () => ({ id: 'ext-id' }) as never,
-    getYomitanParserWindow: () => parserWindow as never,
-    setYomitanParserWindow: () => undefined,
-    getYomitanParserReadyPromise: () => null,
-    setYomitanParserReadyPromise: () => undefined,
-    getYomitanParserInitPromise: () => null,
-    setYomitanParserInitPromise: () => undefined,
-    createYomitanExtensionWindow: options?.createYomitanExtensionWindow as never,
-  };
-}
-
-function createYomitanScriptSandbox(handler: (action: string, params: unknown) => unknown) {
-  return {
-    chrome: {
-      runtime: {
-        lastError: null,
-        sendMessage: (
-          payload: { action?: string; params?: unknown },
-          callback: (response: { result?: unknown; error?: { message?: string } }) => void,
-        ) => {
-          try {
-            callback({ result: handler(payload.action ?? '', payload.params) });
-          } catch (error) {
-            callback({ error: { message: (error as Error).message } });
-          }
-        },
-      },
-    },
-    Array,
-    Error,
-    JSON,
-    Map,
-    Math,
-    Number,
-    Object,
-    Promise,
-    RegExp,
-    Set,
-    String,
-  };
-}
-
-async function runInjectedYomitanScript(
-  script: string,
-  handler: (action: string, params: unknown) => unknown,
-): Promise<unknown> {
-  return await vm.runInNewContext(script, createYomitanScriptSandbox(handler));
-}
-
-// Persistent page context shared across executeJavaScript calls, matching the
-// real parser window: the scan runtime is installed once via
-// globalThis.__subminerYomitanScan and per-line calls reuse it (and its
-// cross-line termsFind cache).
-function createPersistentYomitanScriptRunner(
-  handler: (action: string, params: unknown) => unknown,
-): (script: string) => Promise<unknown> {
-  const context = vm.createContext(createYomitanScriptSandbox(handler));
-  return async (script: string) => await vm.runInContext(script, context);
-}
-
-// Deps whose parser window executes every injected script (profile metadata,
-// scan runtime install, per-line scan calls, parseText fallback) inside one
-// persistent vm context, dispatching backend actions to `handler`.
-function createScanDeps(
-  handler: (action: string, params: unknown) => unknown,
-  options?: { onScript?: (script: string) => void },
-) {
-  const runScript = createPersistentYomitanScriptRunner(handler);
-  return createDeps(async (script) => {
-    options?.onScript?.(script);
-    return await runScript(script);
-  });
-}
-
-function countTermsFindLookups(lookups: string[], prefix: string): number {
-  return lookups.filter((lookupText) => lookupText.startsWith(prefix)).length;
-}
-
-// Backend stub for the greedy name pre-pass: one character name (ミナト) in a
-// line of ordinary words, with the SubMiner character dictionary enabled.
-const NAME_SCAN_WORDS: Array<[string, string, string, boolean]> = [
-  ['ミナト', 'ミナト', 'みなと', true],
-  ['は', 'は', 'は', false],
-  ['まだ', 'まだ', 'まだ', false],
-  ['学校', '学校', 'がっこう', false],
-  ['に', 'に', 'に', false],
-  ['いない', 'いる', 'いる', false],
-];
-
-function createNameScanDeps(
-  lookups: string[],
-  words: Array<[string, string, string, boolean]> = NAME_SCAN_WORDS,
-) {
-  return createScanDeps((action, params) => {
-    if (action === 'optionsGetFull') {
-      return {
-        profileCurrent: 0,
-        profiles: [
-          {
-            options: {
-              scanning: { length: 40 },
-              dictionaries: [
-                { name: 'JMdict', enabled: true, id: 0 },
-                {
-                  name: 'SubMiner Character Dictionary (AniList 1)',
-                  enabled: true,
-                  id: 1,
-                },
-              ],
-            },
-          },
-        ],
-      };
-    }
-    if (action === 'getDictionaryInfo') {
-      return [];
-    }
-    if (action !== 'termsFind') {
-      throw new Error(`unexpected action: ${action}`);
-    }
-    const text = (params as { text?: string } | undefined)?.text ?? '';
-    lookups.push(text);
-    for (const [surface, term, reading, isName] of words) {
-      if (text.startsWith(surface)) {
-        return {
-          originalTextLength: surface.length,
-          dictionaryEntries: [
-            {
-              headwords: [
-                {
-                  term,
-                  reading,
-                  sources: [{ originalText: surface, isPrimary: true, matchType: 'exact' }],
-                },
-              ],
-              definitions: [
-                { dictionary: isName ? 'SubMiner Character Dictionary (AniList 1)' : 'JMdict' },
-              ],
-            },
-          ],
-        };
-      }
-    }
-    return { originalTextLength: 0, dictionaryEntries: [] };
-  });
-}
-
-const NAME_SCAN_LINE = 'ミナトはまだ学校にいない';
-
-test('requestYomitanScanTokens skips name pre-pass lookups where no candidate name can start', async () => {
-  const exhaustiveLookups: string[] = [];
-  const exhaustive = await requestYomitanScanTokens(
-    NAME_SCAN_LINE,
-    createNameScanDeps(exhaustiveLookups),
-    { error: () => undefined },
-    { includeNameMatchMetadata: true },
-  );
-
-  const prefilteredLookups: string[] = [];
-  const prefiltered = await requestYomitanScanTokens(
-    NAME_SCAN_LINE,
-    createNameScanDeps(prefilteredLookups),
-    { error: () => undefined },
-    {
-      includeNameMatchMetadata: true,
-      currentCharacterDictionaryMediaId: 1,
-      // Terms and readings the generated dictionary exposes for this media.
-      nameCandidates: { key: 'media-1', forms: ['ミナト', 'みなと'] },
-    },
-  );
-
-  // Same tokenization, including the name match, with fewer round trips.
-  assert.deepEqual(prefiltered, exhaustive);
-  assert.equal(prefiltered?.[0]?.surface, 'ミナト');
-  assert.equal(prefiltered?.[0]?.isNameMatch, true);
-  assert.ok(
-    prefilteredLookups.length < exhaustiveLookups.length,
-    `expected fewer lookups with candidates (${prefilteredLookups.length} vs ${exhaustiveLookups.length})`,
-  );
-  // Mid-token positions are exactly what the pre-pass used to probe (a name can
-  // start mid-token); with candidates they cost nothing, while the main walk's
-  // own token-start lookups are unaffected.
-  assert.ok(countTermsFindLookups(exhaustiveLookups, '校に') > 0);
-  assert.equal(countTermsFindLookups(prefilteredLookups, '校に'), 0);
-});
-
-test('requestYomitanScanTokens matches a katakana name from its kana-normalized candidate form', async () => {
-  const lookups: string[] = [];
-  const result = await requestYomitanScanTokens(
-    NAME_SCAN_LINE,
-    createNameScanDeps(lookups),
-    { error: () => undefined },
-    {
-      includeNameMatchMetadata: true,
-      currentCharacterDictionaryMediaId: 1,
-      // Only the hiragana reading is listed; the katakana surface in the line
-      // must still be found through kana normalization.
-      nameCandidates: { key: 'media-1', forms: ['みなと'] },
-    },
-  );
-
-  assert.equal(result?.[0]?.surface, 'ミナト');
-  assert.equal(result?.[0]?.isNameMatch, true);
-});
-
-// Kana normalization does not fold halfwidth katakana, so a name written that
-// way can never prefix-match a candidate form; the pre-pass has a bypass for
-// those positions, which only runs if they count as Japanese in the first place.
-// The generic word here reaches into the name, so only a pre-pass reservation
-// can keep the name whole.
-const HALFWIDTH_NAME_SCAN_WORDS: Array<[string, string, string, boolean]> = [
-  ['まだﾐ', 'まだミ', 'まだみ', false],
-  ['まだ', 'まだ', 'まだ', false],
-  ['ﾐﾅﾄ', 'ミナト', 'みなと', true],
-];
-
-test('requestYomitanScanTokens probes halfwidth katakana positions during the name pre-pass', async () => {
-  const lookups: string[] = [];
-  const result = await requestYomitanScanTokens(
-    'まだﾐﾅﾄ',
-    createNameScanDeps(lookups, HALFWIDTH_NAME_SCAN_WORDS),
-    { error: () => undefined },
-    {
-      includeNameMatchMetadata: true,
-      currentCharacterDictionaryMediaId: 1,
-      // Fullwidth forms only, as the generated dictionary stores them.
-      nameCandidates: { key: 'media-1', forms: ['ミナト', 'みなと'] },
-    },
-  );
-
-  assert.equal(countTermsFindLookups(lookups, 'ﾐﾅﾄ'), 1);
-  assert.deepEqual(
-    result?.map((token) => token.surface),
-    ['まだ', 'ﾐﾅﾄ'],
-  );
-  assert.equal(result?.[1]?.isNameMatch, true);
-  // The reading is written the way the fullwidth katakana path writes it
-  // (surface spelling, fullwidth): halfwidth kana is not kana to the known-word
-  // and frequency code downstream, and an empty reading there disables the
-  // reading fallback entirely.
-  assert.equal(result?.[1]?.reading, 'ミナト');
-  assert.equal(result?.[1]?.headwordReading, 'みなと');
-});
-
-test('halfwidth voiced kana compose into the reading instead of leaving a stray mark', async () => {
-  const lookups: string[] = [];
-  const result = await requestYomitanScanTokens(
-    'ｶﾞｸ ﾊﾟﾝ',
-    createNameScanDeps(lookups, [
-      ['ｶﾞｸ', 'ガク', 'がく', false],
-      ['ﾊﾟﾝ', 'パン', 'ぱん', false],
-    ]),
-    { error: () => undefined },
-    { includeNameMatchMetadata: true },
-  );
-
-  const readings = (result ?? [])
-    .filter((token) => token.isUnparsedRun !== true)
-    .map((token) => [token.surface, token.reading]);
-  assert.deepEqual(readings, [
-    ['ｶﾞｸ', 'ガク'],
-    ['ﾊﾟﾝ', 'パン'],
-  ]);
-});
-
-test('requestYomitanScanTokens falls back to the exhaustive name scan without candidates', async () => {
-  const withoutLookups: string[] = [];
-  const withoutCandidates = await requestYomitanScanTokens(
-    NAME_SCAN_LINE,
-    createNameScanDeps(withoutLookups),
-    { error: () => undefined },
-    { includeNameMatchMetadata: true, currentCharacterDictionaryMediaId: 1, nameCandidates: null },
-  );
-
-  assert.equal(withoutCandidates?.[0]?.isNameMatch, true);
-  // No candidate list means every Japanese position is probed, as before.
-  assert.ok(countTermsFindLookups(withoutLookups, '校に') > 0);
-});
-
-test('requestYomitanScanTokens reinstalls name candidates when the media changes', async () => {
-  const lookups: string[] = [];
-  const deps = createNameScanDeps(lookups);
-
-  // First media's candidates cannot match this line's name.
-  const otherMedia = await requestYomitanScanTokens(
-    NAME_SCAN_LINE,
-    deps,
-    { error: () => undefined },
-    {
-      includeNameMatchMetadata: true,
-      currentCharacterDictionaryMediaId: 2,
-      nameCandidates: { key: 'media-2', forms: ['カズマ'] },
-    },
-  );
-  assert.equal(otherMedia?.[0]?.isNameMatch, undefined);
-
-  const correctMedia = await requestYomitanScanTokens(
-    NAME_SCAN_LINE,
-    deps,
-    { error: () => undefined },
-    {
-      includeNameMatchMetadata: true,
-      currentCharacterDictionaryMediaId: 1,
-      nameCandidates: { key: 'media-1', forms: ['ミナト'] },
-    },
-  );
-  assert.equal(correctMedia?.[0]?.surface, 'ミナト');
-  assert.equal(correctMedia?.[0]?.isNameMatch, true);
-});
 
 test('syncYomitanDefaultAnkiServer updates default profile server when script reports update', async () => {
   let scriptValue = '';
